requirements.txt
opencv-python
tensorflow
onnxruntime
Pillow
matplotlib